Steam Web Request Bug, Data is acquired but console.log(err) gives,  ("No 'Access-Control-Allow-Origin' header is present on the requested resource.").
